Initial merge of project to github
Hopefully I didn't screw this up. The sourceforge project administrator
has ignored requests to include this source on his pages, and I've had
requests for a more accessible source release. Here it is.
Note that this merge is different to the last released source, largely
to workaround graphics corruption on OS X Lion and later (due to
something odd happening in the calls to SDL_SetGammaRamp.
The game still runs rather slowly; I have run Measurements on it and
the delay seems to be deep in OS X's image handling routines.