Revert and update to fix score and level issues
SDL_image 1.2.7 is the latest version that seems to work with the level
and score readouts. Due to the various attempts to diagnose this flaw,
the source can now be reverted back to the earlier code and, with
SDL_SetGammaRamp commented out, we're good for Lion and Mountain Lion.
Thanks to the AROS port, the confirmation that the main code was fine
was enough to direct the efforts to understand the issue.